Create spinner for use within profile component
As the current user's profile is being retrieved, a spinner is placed in the UI until the data is back and the store is updated. The user's profile data will then be rendered. This is to ensure that the app does not try to render profile data which it has not yet received from the backend