Render current user's name on profile page in welcome message
The current user's name is dynamically outputted depending on who is logged in. If getCurrentProfile has not returned, the spinner will show. If it is returned, the 'profile' and 'loading' properties of the profile slice of the store will be updated and therefore the welcome message with the logged in user's username will render