Render different text based on the username in the URL
If a user clicks on 'View Profile' on one of the profiles in the beauticians page, the username will automatically be valid as the details for the selected profile was retrived from the database to begin with. Therefore it will always render the user's profile via the SingleProfile component. However, if the client goes directly to /profile/{nonexistent-username}, it will display the message 'This user does not exist'. This is because the SingleProfile component will render, immediately call getProfileByUsername, which will fail and dispatch PROFILE_ERROR. The component will then do a check on the store and only if the profile is empty, will the above error message be rendered to the user