BUG FIX - Remove dispatch of CLEAR_PROFILE action upon logout
If we are on the route /profile/{valid-username} and hit logout, the profile state is removed from the store and the and an error message will appear stating that this user does not exist, which is untrue. This is because the component renders information based on the profile state. This change means that the profile state will remain and the same profile will continue to be rendered accordingly