Add extra server-side validation for social media profiles
If a user enters a website for a social media profile and 'http' or 'https' is not present, the website is not treated independently and is instead appended to the end of the base URL, for example 'localhost:3000/instagram.com/...'. Now the server checks to see if 'https://' exists and if not, it will be added before what was entered. The redirect then works as it should